Add debugger helper to build tasks
- Also extract ProcessResourceFiles class to allow reuse on resource file generator
