+++ source/Tools.BuildTasks/TasksConstants.cs
@@ -1,0 +1,13 @@
+﻿//
+// Copyright (c) 2017 The nanoFramework project contributors
+// Portions Copyright (c) Microsoft Corporation.  All rights reserved.
+// See LICENSE file in the project root for full license information.
+//
+
+namespace nanoFramework.Tools
+{
+    internal class TasksConstants
+    {
+        public const string BuildTaskDebugVar = "NFBUILD_TASKS_DEBUG";
+    }
+}
