+++ source/Tools.BuildTasks/Tools.BuildTasks.csproj
@@ -49,10 +49,13 @@
   <ItemGroup>
     <Compile Include="GenerateBinaryOutputTask.cs" />
     <Compile Include="MetaDataProcessorTask.cs" />
+    <Compile Include="ProcessResourceFiles.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Extensions\CommandLineBuilderExtensions.cs" />
     <Compile Include="GenerateNanoResourceTask.cs" />
     <Compile Include="ResolveRuntimeDependenciesTask.cs" />
+    <Compile Include="TasksConstants.cs" />
+    <Compile Include="Utilities\DebuggerHelper.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="nanoFrameworkBuild.targets" />
